create pat_contract_spreadseet
this program creates the contract breakdown spreadsheets pat and nick fusee used to manually create, as of how the salary cap and contract tables were structured in March 2020. this will probably have to be updated in order to be run again with all the updates made to the cap and contract tables.